requirements.txt
requests
tabulate
python-dateutil